% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr.data.R
\docType{data}
\name{scr.data}
\alias{scr.data}
\title{Spatial Capture-Recapture test/example data}
\format{A list containing examples of data used in and generated by SCR methods.}
\usage{
data(scr.data)
}
\description{
A dataset containing examples of simple SCR capture histories,
an acoustic capture history, trap co-ordinates, and mask point co-ordinates.
}
\examples{
data(scr.data)
caps = scr.sim(5, 75, scr.data$traps)
}
\keyword{datasets}
